+++ CHANGELOG.md
@@ -8,6 +8,9 @@ format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - In the Danger Deck activity, add support for removing cards from the
   deck, needed for the Mean Streets "The Deadly Smell of Disease"
   mission ([issue #19](https://github.com/kuhrusty/MorbadScorepad/issues/19)).
+- In the Skill List activity, add support for filtering skills by the
+  amount of XP you have
+  ([issue #18](https://github.com/kuhrusty/MorbadScorepad/issues/18)).
 
 #### Changed
 - In the Danger Deck activity, fiddle with the content & position of the